chore(module): rename to modules
