+++ package.json
@@ -15,13 +15,16 @@
   "dependencies": {
     "eslint-config-airbnb-typescript": "^18.0.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-page-scroller": "^3.0.1",
+    "react-youtube": "^10.1.0"
   },
   "devDependencies": {
     "@testing-library/react": "^16.0.0",
     "@types/jest": "^29.5.12",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
+    "@types/youtube-player": "^5.5.11",
     "@typescript-eslint/eslint-plugin": "^7.4.0",
     "@typescript-eslint/parser": "^7.4.0",
     "@vitejs/plugin-react": "^4.2.1",
