+++ src/modules/Josephus/KillerInput/KillerInput.tsx
