+++ src/modules/Josephus/KillerInput/index.ts
