+++ src/modules/Josephus/index.ts
