+++ vite.config.ts
@@ -1,7 +1,13 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-})
+  define: {
+    // Some libraries use the global object, even though it doesn't exist in the browser.
+    // Alternatively, we could add `<script>window.global = window;</script>` to index.html.
+    // https://github.com/vitejs/vite/discussions/5912
+    global: {},
+  },
+});
